Seperate core logic from RestGetAllMcis, RestDelAllMcis, ...
